refactor: modernised enrollment endpoint
The enrollment endpoint was using deprecated Registry values which have all been replaced which the new way of configuring the MDM server. The WAP provisioning payload is also now minified by the code which makes the payload much easier to read and work on. Unneeded logging was removed from the endpoint as well.