feat: basic working manage endpoint
This endpoint is only designed to give a valid response when the device "checks in". This results in the Sync showing up as successful in the devices setting. The endpoint has 2 hardcoded responses one for a normal MDM "checkin" and one for an AzureAD "checkin" which requires an extra "Status" in the response.