fix: working AAD join
Working AD join by changing the cert store that the signed certificate is stored to on the device. The MDM policy endpoint has been redone to be as minimal as possible (as its response is being ignored by the client). Added an extra warning to the startup messages about how MDM security is important. Removed all req/res logging.